chore(versioning): add API versioning
- add api versioning
- fix the specs accordingly

[Finishes #154478101]
