+++ app/constraints/api_constraint.rb
@@ -1,0 +1,10 @@
+class ApiConstraint
+  def initialize(options)
+    @version = options[:version]
+    @default = options[:default]
+  end
+
+  def matches?(request)
+    @default || request.headers.fetch(:accept).include?("version=#{@version}")
+  end
+end
