+++ spec/requests/users/get_user_status_posts_spec.rb
@@ -1,13 +1,13 @@
 require "rails_helper"
 
-RSpec.describe UsersController, type: :request do
+RSpec.describe V1::UsersController, type: :request do
   let!(:user) { create :user }
   let!(:status_post) { create :status_post, posted_by: user.id }
   let(:json) { JSON.parse(response.body) }
   let(:headers) { set_header(user.id) }
 
-  describe "GET /api/v1/users/:id/status_posts" do
-    before { get "/api/v1/users/#{user.id}/status_posts", headers: headers }
+  describe "GET /users/:id/status_posts" do
+    before { get "/users/#{user.id}/status_posts", headers: headers }
 
     it "returns status posts made by the user" do
       expect(json).not_to be_empty
